requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
choices==0.1
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2.12
django-admin-better-image==0.2
django-allauth==0.41.0
django-appconf==1.0.4
django-ckeditor==5.9.0
django-image==0.1.0
django-image-optimizer==1.0.1
django-imagekit==4.0.2
django-import-export==2.0.2
django-js-asset==1.2.2
django-optimized-image==0.3.0
django-qr-code==1.1.0
django-qr-creator==0.1.5
django-qrcode==0.3
django-rest-framework==0.1.0
django-select-multiple-field==0.4.2
django-suit==0.2.28
django-suit-redactor-django2==0.5.1
django-watermark==0.1.8
django-watermarker==1.3.3
djangorestframework==3.11.0
et-xmlfile==1.0.1
httpie==2.0.0
idna==2.9
jdcal==1.4.1
MarkupPy==1.14
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.3
pilkit==2.0
Pillow==7.0.0
psycopg2-binary==2.8.4
Pygments==2.6.1
python-decouple==3.3
python-resize-image==1.1.19
python3-openid==3.1.0
pytz==2019.3
PyYAML==5.3.1
qrcode==5.1
reportlab==3.5.42
requests==2.23.0
requests-oauthlib==1.3.0
reverse==0.1.0
six==1.14.0
sqlparse==0.3.1
tablib==1.1.0
tinify==1.5.1
urllib3==1.25.8
xlrd==1.2.0
xlwt==1.3.0
